move experiment logic out of session, standardize timestamp in util
