+++ rl/session.py
@@ -1,7 +1,6 @@
 import gym
 import json
 import multiprocessing as mp
-from datetime import datetime
 from functools import partial
 from keras import backend as K
 from rl.spec import game_specs
@@ -70,101 +69,3 @@ class Session(object):
 
         K.clear_session()  # manual gc to fix TF issue 3388
         return sys_vars
-
-
-def experiment_analytics(data):
-    '''
-    helper: define the performance metric
-    given data from an experiment
-    '''
-    sys_vars_array = data['sys_vars_array']
-    mean_r_array = [sys_vars['mean_rewards'] for sys_vars in sys_vars_array]
-    metrics = {
-        'experiment_mean': np.mean(mean_r_array),
-        'experiment_std': np.std(mean_r_array),
-    }
-    return metrics
-
-
-def save_experiment_data(data_grid):
-    '''
-    log the entire experiment data grid from inside run()
-    '''
-    # sort data, best first
-    data_grid.sort(
-        key=lambda data: data['metrics']['experiment_mean'],
-        reverse=True)
-    timestamp = '{:%Y-%m-%d_%H%M%S}'.format(datetime.now())
-    filename = './data/{}_{}_{}_{}_{}.json'.format(
-        data_grid[0]['sess_spec']['problem'],
-        data_grid[0]['sess_spec']['Agent'],
-        data_grid[0]['sess_spec']['Memory'],
-        data_grid[0]['sess_spec']['Policy'],
-        timestamp
-    )
-    with open(filename, 'w') as f:
-        json.dump(data_grid, f, indent=2, sort_keys=True)
-    logger.info('Experiment complete, written to data/')
-
-
-def run_single_exp(sess_spec, data_grid, times=1):
-    '''
-    helper: run a experiment for Session
-    a number of times times given a sess_spec from gym_specs
-    '''
-    start_time = datetime.now().isoformat()
-    sess = Session(problem=sess_spec['problem'],
-                   Agent=sess_spec['Agent'],
-                   Memory=sess_spec['Memory'],
-                   Policy=sess_spec['Policy'],
-                   param=sess_spec['param'])
-    sys_vars_array = [sess.run() for i in range(times)]
-    end_time = datetime.now().isoformat()
-    data = {  # experiment data
-        'start_time': start_time,
-        'sess_spec': stringify_param(sess_spec),
-        'sys_vars_array': sys_vars_array,
-        'metrics': None,
-        'end_time': end_time,
-    }
-    data.update({'metrics': experiment_analytics(data)})
-    # progressive update of data_grid, write when an exp is done
-    data_grid.append(data)
-    save_experiment_data(data_grid)
-    return data
-
-
-def run(sess_name, run_param_selection=False, times=1, line_search=True):
-    '''
-    primary method:
-    run the experiment (single or multiple)
-    specifying if this should be a param_selection run
-    and run each for a number of times
-    calls run_single_exp internally
-    and employs parallelism whenever possible
-    '''
-    sess_spec = game_specs.get(sess_name)
-    data_grid = []
-
-    if run_param_selection:
-        if line_search:
-            param_grid = param_line_search(
-            sess_spec['param'], sess_spec['param_range'])
-        else:
-            param_grid = param_product(
-                sess_spec['param'], sess_spec['param_range'])
-        sess_spec_grid = [{
-            'problem': sess_spec['problem'],
-            'Agent': sess_spec['Agent'],
-            'Memory': sess_spec['Memory'],
-            'Policy': sess_spec['Policy'],
-            'param': param,
-        } for param in param_grid]
-        p = mp.Pool(mp.cpu_count())
-        list(p.map(
-            partial(run_single_exp, data_grid=data_grid, times=times),
-            sess_spec_grid))
-    else:
-        run_single_exp(sess_spec, data_grid=data_grid, times=times)
-
-    return data_grid
